app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
openai
prometheus-fastapi-instrumentator
uuid
pydantic